fix: remove custom config parsing logic
use standard config parsing logic in repolinter, which includes
additional functionality like extending config files.

Config file paths are now interpreted as relative to the repolinter
working directory rather than the root of the repository.  This
difference can be seen in the 'runs in a custom directory' test.  This
only breaks if you are running repolinter in a custom directory AND
using a custom config file, which we don't think anyone is actually
doing.

Fixes #4